custom_components/ha_alarm_clock/requirements.txt
voluptuous>=0.13.1
mutagen>=1.47.0
pytest
pytest-homeassistant-custom-component